% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barbell.R
\docType{data}
\name{GeomBarbell}
\alias{GeomBarbell}
\title{Barbell Plot Geom}
\format{
An object of class \code{GeomBarbell} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomBarbell
}
\description{
Barbell Plot Geom
}
\references{
https://ggplot2-book.org/extensions.html#combining-multiple-geoms
}
\keyword{internal}
