% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\docType{data}
\name{GeomForest}
\alias{GeomForest}
\title{Forest Plot Geom}
\format{
An object of class \code{GeomForest} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomForest
}
\description{
Forest Plot Geom
}
\keyword{internal}
